вынес форму добавления комментария в отдельный компонент
